test: Debug - Why Trove macros don't work in Scittle
KEY DISCOVERY:
log! is NOT in the public symbols of taoensso.trove!
Only *log-fn* is public.

This explains why:
- (trove/log! ...) fails - log! is not a regular symbol
- (require '[... :refer [log!]]) fails - can't refer to non-public symbols
- Macros are handled differently than functions in SCI

The issue is NOT that macros don't work in Scittle.
The issue is that Trove's log! macro is not properly exported
for use in SCI/Scittle context.

This confirms our wrapper approach is necessary:
- We can't access Trove's macros directly
- We need to use the console backend (functions)
- Our wrapper provides the interface

Files:
- test-trove-unqualified.html - Multi-script test
- test-trove-unqualified-single.html - Single-script test
- test-trove-debug.html - Debug to find why it fails